requirements.txt
streamlit
langchain
langchain-community
langchain-openai
unstructured[all-docs]
faiss-cpu
sentence-transformers
pandas
openai
